backend/requirements.txt
fastapi
pydantic
sqlalchemy
databases
email-validator
requests
python-dotenv